refactor: cria arquivos logic e consts no modulo WithFilters
