+++ src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/modules/dashboards/modules/WithFilters/consts.ts
@@ -1,0 +1,23 @@
+export const clearFiltersArgs = {
+    iconActive: "https://media.graphassets.com/1EOGJH6TvSMqTrjigY1g",
+    iconInactive: "https://media.graphassets.com/x37RkcUrTH6G50ganj9d",
+    label: "Limpar todos os filtros",
+};
+
+//Descricão dos filtros para impressão
+export const filtersLabels = {
+    conditionIdentifiedBy: "Tipo de Diagnóstico",
+    visitantCommunityHealthWorker: "Prof. Responsável",
+    patientStatus: "Situação",
+    patientAgeRange: 'Faixa Etária',
+};
+export const referenceOrder = [
+    "Menos de 17 anos",
+    "Entre 18 e 24 anos",
+    "Entre 25 e 34 anos",
+    "Entre 35 e 44 anos",
+    "Entre 45 e 54 anos",
+    "Entre 55 e 65 anos",
+    "Mais de 65 anos"
+];
+
