--- src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/modules/dashboards/modules/WithFilters/modules/clearFiltersArgs.ts
@@ -1,5 +1,0 @@
-export const clearFiltersArgs = {
-    iconActive: "https://media.graphassets.com/1EOGJH6TvSMqTrjigY1g",
-    iconInactive: "https://media.graphassets.com/x37RkcUrTH6G50ganj9d",
-    label: "Limpar todos os filtros",
-};
