--- src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/modules/dashboards/modules/WithFilters/modules/filtersLabels.ts
@@ -1,7 +1,0 @@
-//Descricão dos filtros para impressão
-export const filtersLabels = {
-    conditionIdentifiedBy: "Tipo de Diagnóstico",
-    visitantCommunityHealthWorker: "Prof. Responsável",
-    patientStatus: "Situação",
-    patientAgeRange: 'Faixa Etária',
-};
